% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_n.R
\name{add_n}
\alias{add_n}
\title{A function to the number of individual entries to a vector.}
\usage{
add_n(v)
}
\arguments{
\item{l}{A vector.}
}
\description{
To be used with ggplot2::labeller().
}
\examples{
add_n(iris$Species)

}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
