% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_l_vec.R
\name{bind_l_vec}
\alias{bind_l_vec}
\title{Bind a list of vectors of different lengths.}
\usage{
bind_l_vec(l)
}
\arguments{
\item{l}{A list of vectors.}
}
\description{
Bind a list of vectors of different lengths.
}
\examples{
l = list(col1 = c('A', 'B'),
col2  =c('A'))
bind_l_vec(l)

}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
