% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casconv.R
\name{casconv}
\alias{casconv}
\title{Function to convert between hypen and hyphen-less CAS strings.}
\usage{
casconv(x, hyphen = TRUE, na_999 = FALSE)
}
\arguments{
\item{x}{A vector of CAS numbers, either with hyphen (e.g. 50-00-0) or
without hyphen (e.g. 50000)}

\item{hyphen}{Should the hyphen or the hyphen-less CAS string be returned?}

\item{na_999}{Should NAs be replaced with 999-99-9 or 999999 ?}
}
\description{
Function to convert between hypen and hyphen-less CAS strings.
}
\examples{
vec = c('50-00-0', '1071836', '1071-83-6')
casconv(vec)
casconv(vec, hyphen = FALSE)
casconv(vec, hyphen = FALSE, na_999 = TRUE)
}
