% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{chck_dupl}
\alias{chck_dupl}
\title{Check if duplicated values are in columns of data.frame.}
\usage{
chck_dupl(dat, col = NULL)
}
\arguments{
\item{dat}{A data.frame.}

\item{col}{A character vector specifying the columns.}
}
\description{
Check if duplicated values are in columns of data.frame.
}
