% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{chck_na}
\alias{chck_na}
\title{Check if NA values are in columns of data.frame.}
\usage{
chck_na(dat, col = NULL)
}
\arguments{
\item{dat}{A data.frame.}

\item{col}{A character vector specifying the columns.}
}
\description{
Check if NA values are in columns of data.frame.
}
