% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_read.R
\name{cl_read}
\alias{cl_read}
\title{Function to load lookup tables from the andschar/chem-lookup GitHub repository}
\usage{
cl_read(table = "cl_identifiers", token = NULL)
}
\arguments{
\item{table}{Which chem-lookup table should be loaded? Can be one of \code{
'cl_identifiers', 'cl_class', 'cl_role' 'cl_pesticide', 'cl_identifiers_meta'
}}

\item{token}{Provide a GitHub personal access token (PAT) here or store it in
~/.Renviron}
}
\description{
Function to load lookup tables from the andschar/chem-lookup GitHub repository
}
\examples{
cl_read('cl_identifiers')

}
