% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_workspace.R
\name{clean_workspace}
\alias{clean_workspace}
\title{Function to clean workspace appart from some variables}
\usage{
clean_workspace(keep = NULL, envir = .GlobalEnv)
}
\arguments{
\item{keep}{Objects to not be removed from the workspace. Provide as 
character vector}

\item{envir}{Envrionment to be cleaned. Deafault: .GlobalEnv}
}
\description{
Function to clean workspace appart from some variables
}
\examples{
clean_workspace()

}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
