% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_labeler.R
\name{cut_labeler}
\alias{cut_labeler}
\title{Function to produce nice Labels for the cut function}
\usage{
cut_labeler(x, round_digits = 0)
}
\arguments{
\item{v}{A numeric vector of \code{cit()} breaks.}
}
\description{
Function to produce nice Labels for the cut function
}
\examples{
cut_labeler(c(5, 10, 15))

}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
