% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_up.R
\name{firstup}
\alias{firstup}
\title{Function to upper-case 1st letter}
\usage{
firstup(x)
}
\arguments{
\item{x}{character; String for which the first letter should be set to upper
case}
}
\description{
Function to upper-case 1st letter
}
\examples{
firstup('andreas')


}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
