% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixweek.R
\name{fixweek}
\alias{fixweek}
\title{Convenience function to turn week 1-9 into 01-09.}
\usage{
fixweek(x)
}
\arguments{
\item{x}{a week number (as string or character).}
}
\description{
Convenience function to turn week 1-9 into 01-09. Also rounds
float values. This is important in as.Date() parsing.
}
\examples{
fixweek(9)

}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
