% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_nice.R
\name{form_nice}
\alias{form_nice}
\title{Function to produce nice numeric labels}
\usage{
form_nice(x, big.mark = ",", scientific = FALSE, drop0trailing = TRUE)
}
\arguments{
\item{x}{A scalar or vector passed on to \code{format()}.}

\item{big.mark}{Same as in \code{format()}.}

\item{scientific}{Same as in \code{format()}.}

\item{drop0trailing}{as in \code{format()}.}

\item{...}{Other arguments passed on to \code{format()}.}
}
\description{
Function to produce nice numeric labels
}
\examples{
x = c(0.0001, 5000, 2.45e5)
form_nice(x)

}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
