% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freadwrite2.R
\name{fwrite2}
\alias{fwrite2}
\title{Convenience data.table::fwrite() wrapper.}
\usage{
fwrite2(..., sep = "\\t", quote = FALSE)
}
\arguments{
\item{...}{Any parameters for fwrite().}

\item{sep}{Separator as in fwrite().}

\item{quote}{Quote as in fwrite().}
}
\description{
Convenience data.table::fwrite() wrapper.
}
\examples{
fwrite2(iris, file.path(tempdir(), 'iris.tsv'),
        sep = '|')

}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
