% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_dummy.R
\name{gg_dummy}
\alias{gg_dummy}
\title{Function to produce a dummy ggplot}
\usage{
gg_dummy()
}
\description{
Function to produce a dummy ggplot
}
\examples{
gg = gg_dummy() 

}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
