% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{\%llike\%}
\alias{\%llike\%}
\title{Helper functions to apply grepl() to list column elements}
\usage{
lhs \%llike\% rhs
}
\arguments{
\item{lhs}{Character vector.}

\item{rhs}{Character vector that should be looked up in x.}
}
\description{
Helper functions to apply grepl() to list column elements
}
\examples{
require(data.table)
dt = data.table(a = c('R1', 'R2', 'R3', 'R4', 'R5'),
                b = list(c('A', 'B'),
                c('A'),
                list('A', 'B', 'C'),
                list(list(c('B', 'C'))),
                'B'))
dt[ a \%llike\% '[0-3]' ]

}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
