% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksource.R
\name{ksource}
\alias{ksource}
\title{A function to source .Rmd files.}
\usage{
ksource(x, ...)
}
\arguments{
\item{x}{Path to a .Rmd file.}

\item{...}{Arguments to pass on.}
}
\description{
A function to source .Rmd files.
}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
