% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_right.R
\name{left}
\alias{left}
\title{Function to trim n leading characters from string.}
\usage{
left(string, n)
}
\arguments{
\item{string}{String of vector of strings.}

\item{n}{Number of characters to be trimmed}
}
\description{
Inspired by PosgreSQL
\url{https://www.postgresql.org/docs/current/functions-string.html}.
}
\examples{
v = c('1234567', 'abcdefg')
left(v, 2)

}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
