% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_msg}
\alias{log_msg}
\title{Function to log a msg to a file}
\usage{
log_msg(msg = "script run.", file = "script.log")
}
\arguments{
\item{msg}{Message to log.}

\item{file}{file to log to. Stored in the root directory.}
}
\description{
Function to log a msg to a file
}
\examples{
\donttest{
  # connection to database required
  log_msg('Test')
}

}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
