% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_number.R
\name{max_number}
\alias{max_number}
\title{Function to replace larger values with a threshold value.}
\usage{
max_number(x, y)
}
\arguments{
\item{x}{A numeric vector or scalar.}

\item{y}{A numeric threshold value that serves as a cutoff.}
}
\description{
Function to replace larger values with a threshold value.
}
\examples{
max_number(100, 10)

}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
