% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_number.R
\name{min_number}
\alias{min_number}
\title{Function to replace smaller values with a threshold value.}
\usage{
min_number(x, y)
}
\arguments{
\item{x}{A numeric vector or scalar.}

\item{y}{A numeric threshold value that serves as a cutoff.}
}
\description{
Function to replace smaller values with a threshold value.
}
\examples{
min_number(3, 10)

}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
