% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namer.R
\name{namer}
\alias{namer}
\title{Function to create file name (path) string.}
\usage{
namer(
  dir = NULL,
  prequel = NULL,
  arg = NULL,
  sep = getOption("sep"),
  ext = NULL
)
}
\arguments{
\item{dir}{A directory to which the path should point to.}

\item{prequel}{Prequel that is appended to the file name (e.g. '~/project').}

\item{arg}{A named list or vector of arguments that are used for creating
the file name. Once created, they can be extracted with \code{denamer()}.}

\item{sep}{String that separates names and entries of the arg
(Default: <_>).}

\item{ext}{Should a file extension be added at the end of the string?
(e.g. .png)?}
}
\description{
A file path is created in the form <dir>/<prequel><arg><ext>.
}
\examples{
namer(dir = tempdir(),
      prequel = 'test-prequel',
      arg = list(weight = 12,
                 color = 'white'),
      sep = '_',
      ext = '.png')

}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
