% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth_element.R
\name{nth_element}
\alias{nth_element}
\title{Function to return every nth element of a vector}
\usage{
nth_element(x, n = 2)
}
\arguments{
\item{v}{A character vector.}
}
\description{
Useful fo reducing entry overload for discrete axis on a plot.
Helper function.
}
\examples{
x = row.names(mtcars)
nth_element(x, 3)

}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
