% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdepend.R
\name{rdepend}
\alias{rdepend}
\title{Function to check what packages are used in multiple scripts in a directory}
\usage{
rdepend(path = ".")
}
\arguments{
\item{path}{Specify root directory.}
}
\description{
Function to check what packages are used in multiple scripts in a directory
}
\examples{
\donttest{
  # Not certain if a root directory exists
  rdepend('.')
}

}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
