% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_io.R
\name{read_query}
\alias{read_query}
\title{Function to read from a database.}
\usage{
read_query(query = NULL, drv, dbname, cred_file = NULL, ..., quiet = FALSE)
}
\arguments{
\item{query}{A query string.}

\item{drv}{A database driver.}

\item{cred_file}{An .R file containing the necessary credentials to connect
to a different databases (e.g. PostgreSQL). Can contain variables: user,
host, port, password, dbname.}

\item{...}{Arguments passed on to \code{DBI::dbConnect()}.}

\item{quiet}{Quiet function call?}
}
\description{
Function to read from a database.
}
\examples{
\donttest{
# connection to database required
read_query(query = "SELECT * FROM schema.table LIMIT 1;",
           cred_file = cred.R)
}

}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
