% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readChar2.R
\name{readChar2}
\alias{readChar2}
\title{Wrapper around \code{readChar()}}
\usage{
readChar2(con)
}
\arguments{
\item{con}{A connection object, or a character string naming a file, or a raw vector.
Same as in \code{readChar()}.}
}
\description{
Wrapper around \code{readChar()}
}
\examples{
tmp = tempfile()
writeLines('Hello\nDarkness\nFriend', tmp)
readChar2(tmp)

}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
