% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_reorder.R
\name{reorder_within}
\alias{reorder_within}
\title{Helper function to reorder ggplot facets.}
\usage{
reorder_within(x, by, within, fun = mean, sep = "___", ...)
}
\arguments{
\item{x}{The column to reorder.}

\item{by}{The column with which \code{x} should be ordered.}

\item{fun}{Passed on to stats::reorder().}

\item{sep}{Separator string used to concatenate columns.}

\item{...}{Currently not used.}
}
\description{
Taken from \url{https://stackoverflow.com/questions/52214071}.
}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
