% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_reorder.R
\name{scale_x_reordered}
\alias{scale_x_reordered}
\title{Helper function to reorder ggplot facets.}
\usage{
scale_x_reordered(..., sep = "___")
}
\arguments{
\item{...}{Parameters passed to \code{ggplot2::scale_x_discrete()}.}

\item{sep}{Separator string to split columns.}
}
\description{
Taken from \url{https://stackoverflow.com/questions/52214071}.
}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
