% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{size}
\alias{size}
\title{Function to print R object pretty size}
\usage{
size(obj)
}
\arguments{
\item{obj}{An R object.}
}
\description{
Function to print R object pretty size
}
\examples{
size(iris)

}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
