% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subchunkify.R
\name{subchunkify}
\alias{subchunkify}
\title{Function to create different heights and widths in RMarkdown Code Chunks which are Created automatically}
\usage{
subchunkify(
  g,
  fig.height = 7,
  fig.width = 5,
  echo = FALSE,
  message = FALSE,
  warning = FALSE
)
}
\arguments{
\item{g}{A plot.}

\item{fig.height}{The Figure height as in an RMarkdown document.}

\item{fig.width}{The Figure width as in an RMarkdown document.}

\item{echo}{As in an RMarkdown document.}

\item{message}{As in an RMarkdown document.}

\item{warning}{As in an RMarkdown document.}
}
\description{
Taken from StackOverflow.
}
\examples{
subchunkify(plot(iris, Sepal.Length ~ Sepal.Width))

}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
