% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_theme.R
\name{theme_as}
\alias{theme_as}
\title{Custom ggplot2 theme}
\usage{
theme_as(base_size = 12, base_family = "Open Sans", text_size = 14)
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}

\item{text_size}{Size of text.}
}
\description{
Custom ggplot2 theme
}
\examples{
library(ggplot2)
p = ggplot(mtcars) + 
 geom_point(aes(x = wt, y = mpg, 
colour = factor(gear))) + facet_wrap(~am)
p
p + theme_as()

}
