% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toupper_all.R
\name{toupper_all}
\alias{toupper_all}
\title{Function for Capitalize First Letters}
\usage{
toupper_all(x)
}
\arguments{
\item{x}{A string.}
}
\description{
Function for Capitalize First Letters of all Words in a string.
Taken from \code{?toupper()}.
}
\examples{
s = c('hello darkness my old Friend')
toupper_all(s)

}
