% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_seperator.R
\name{val_seperator}
\alias{val_seperator}
\title{Function to separate values at the nth position}
\usage{
val_seperator(v, n = 3)
}
\arguments{
\item{v}{A numeric vector.}

\item{n}{A numeric vector indicating the position to separate.}
}
\description{
Function to separate values at the nth position
}
\examples{

v = 21032262 
val_seperator(v)
val_seperator(v, 2)

}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
