% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_io.R
\name{write_tbl}
\alias{write_tbl}
\title{Functions to write a table to a database.}
\usage{
write_tbl(
  dat = NULL,
  schema = NULL,
  tbl = NULL,
  key = NULL,
  comment_str = NULL,
  overwrite = TRUE,
  drv,
  dbname,
  cred_file = NULL,
  ...,
  quiet = FALSE
)
}
\arguments{
\item{dat}{data.frame, data.table or tibble.}

\item{schema}{database schema.}

\item{tbl}{database table.}

\item{key}{Set a primary key in table?}

\item{comment_str}{Add a comment to table? Recommended.}

\item{overwrite}{Should an existing database table be overwritten?
TRUE (default).}

\item{drv}{A database driver.}

\item{dbname}{The database name.}

\item{cred_file}{An .R file containing the necessary credentials to connect
to a different databases (e.g. PostgreSQL). Can contain variables: user,
host, port, password, dbname.}

\item{...}{Arguments passed on to \code{DBI::dbConnect()}.}

\item{quiet}{Quiet function call?}
}
\description{
Functions to write a table to a database.
}
\examples{
\donttest{
# connection to database required
read_query(query = "SELECT * FROM schema.table LIMIT 1;",
           cred_file = cred.R)
}

}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
