% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_msg}
\alias{log_msg}
\title{Function to log a msg to a file}
\usage{
log_msg(msg = "script run.", dir = NULL, file = "script.log")
}
\arguments{
\item{msg}{Message to log.}

\item{dir}{A directory to write the report to.}

\item{file}{file to log to.}
}
\description{
Function to log a msg to a file
}
\examples{
\donttest{
  # connection to database required
  log_msg('Test')
}

}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
