% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_io.R
\name{send_query}
\alias{send_query}
\title{Function to send to a database.}
\usage{
send_query(
  query = NULL,
  cred_file = NULL,
  user = NULL,
  host = NULL,
  port = NULL,
  password = NULL,
  dbname = NULL,
  quiet = FALSE
)
}
\arguments{
\item{query}{A query string.}

\item{cred_file}{An .R file containing the necessary credentials to connect
to a Postgres database. Can contain variables: user, host, port, password,
dbname.}

\item{user}{Optional user. Overwritten if \code{user} is defined in
cred_file.}

\item{host}{Optional host. Overwritten if \code{host} is defined in
cred_file.}

\item{port}{Optional port Overwritten if \code{port} is defined in
cred_file.}

\item{password}{Optional password Overwritten if \code{password} is defined
in cred_file.}

\item{dbname}{Optional dbname Overwritten if \code{dbname} is defined
in cred_file.}

\item{quiet}{Quiet function call?}
}
\description{
Function to send to a database.
}
\examples{
\donttest{
# connection to database required
read_query(query = "SELECT * FROM schema.table LIMIT 1;",
           cred_file = cred.R)
}

}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
