% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subchunkify.R
\name{subchunkify}
\alias{subchunkify}
\title{Function to create different heights and widths in RMarkdown Code Chunks which are Created automatically}
\usage{
subchunkify(g, fig_height = 7, fig_width = 5)
}
\arguments{
\item{g}{A plot.}

\item{fig_height}{The Figure height as in an RMarkdown document.}

\item{fig_width}{The Figure width as in an RMarkdown document.}
}
\description{
Taken from StackOverflow.
}
\examples{
subchunkify(plot(iris, Sepal.Length ~ Sepal.Width))

}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
