% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_io.R
\name{write_tbl}
\alias{write_tbl}
\title{Functions to write a table to a database.}
\usage{
write_tbl(
  dat = NULL,
  schema = NULL,
  tbl = NULL,
  key = NULL,
  comment_str = NULL,
  overwrite = TRUE,
  cred_file = NULL,
  host = NULL,
  port = NULL,
  dbname = NULL,
  user = NULL,
  password = NULL,
  quiet = FALSE
)
}
\arguments{
\item{dat}{data.frame, data.table or tibble.}

\item{schema}{database schema.}

\item{tbl}{database table.}

\item{key}{Set a primary key in table?}

\item{comment_str}{Add a comment to table? Recommended.}

\item{overwrite}{Should an existing database table be overwritten?
TRUE (default).}

\item{cred_file}{An .R file containing the necessary credentials to connect
to a Postgres database. Can contain variables: user, host, port, password,
dbname.}

\item{host}{Optional host. Overwritten if \code{host} is defined in
cred_file.}

\item{port}{Optional port Overwritten if \code{port} is defined in
cred_file.}

\item{dbname}{Optional dbname Overwritten if \code{dbname} is defined
in cred_file.}

\item{user}{Optional user. Overwritten if \code{user} is defined in
cred_file.}

\item{password}{Optional password Overwritten if \code{password} is defined
in cred_file.}

\item{quiet}{Quiet function call?}
}
\description{
Functions to write a table to a database.
}
\examples{
\donttest{
# connection to database required
read_query(query = "SELECT * FROM schema.table LIMIT 1;",
           cred_file = cred.R)
}

}
\author{
Andreas Scharmueller, \email{andschar@proton.me}
}
